pyproject.toml
[project]
name = "util"
version = "0.1.5"
authors = [{name = "M", email = "refits.minxes0h@icloud.com"}]
dynamic = ["version", "description"]
dependencies = [
    "python-magic",
    "pyyaml"
]

[license]
file = "UNLICENSE"

[tool.flit.sdist]
exclude = ["tests/"]

[tool.poetry]
name = "util"
version = "0.1.5"
description = "General funcs & classes; extending standard libs"
authors = ["M <refits.minxes0h@icloud.com>"]
license = "UNLICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-magic = "^0.4.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"